Split algorithm runs into data-only and brokerage-only steps
Algorithms now run as two stages sharing one code path between the
scheduled runner and the MCP agent:

  step 1  run_algorithm()  - algorithm + data sources, no brokerage
  step 2  place_orders()   - algorithm + brokerage, no data fetching

Prices ride along on the step-1 result so step 2 never touches a data
source, and a stale result is refused rather than sized against old
prices. The only difference between the scheduled and agent flows is
that the agent validates in between.

Position-aware logic (stickiness, risk guards, sizing) moves into
refine(), which takes a PortfolioSnapshot rather than a Brokerage so it
also works for backtests and for accounts with no broker configured.

Consolidation:
- one scoring path per algorithm; all three signal_view overrides and
  both build_*_targets helpers removed
- deleted sync_positions_to_targets, defensive_momentum_portfolio_preview
  and assorted dead helpers (~560 lines net)

New:
- PaperBrokerage for running with no broker configured
- SchwabBrokerage + OAuth session (UNVERIFIED - no approved app yet)
- Schwab quotes wired into the cached provider path; the existing Schwab
  bar fetchers now refresh their token via OAuth instead of a static one

Fixes:
- fast_momentum/invest_spy were double-counting the cash buffer and
  under-investing by ~2%; sizing() now lets an algorithm opt out
- unpriceable symbols no longer abort a whole run
- a rejected order leg no longer aborts the rest of the batch
- Schwab order ids were lost from the 201 Location header

Note: sentiment_scores_from_records and _defensive_momentum_reason were
reconstructed from their call sites after an accidental truncation; they
satisfy the tests but were not recovered from the originals.

+++ src/brokerages/base.py
@@ -11,7 +11,7 @@ class BaseBrokerage(Brokerage):
     def get_account_state(self) -> Dict[str, Any]:
         raise NotImplementedError
 
-    def get_positions(self) -> Dict[str, int]:
+    def get_positions(self) -> Dict[str, float]:
         raise NotImplementedError
 
     def submit_order(self, request: OrderRequest) -> Dict[str, Any]:
